# Make quest invite modal pop up on user sync, challenge modal display before invite modal, and additional party tweaks