+ Added a Boolean `invited` field to the user model that defaults to `false`
+ Set `invited` to true for all group members (other than quest leader) when quest invitation is sent ([code link](https://github.com/HabitRPG/habitrpg/blob/0fa93506f5ea2eb68284486dc0a37bd970c1e7a3/website/src/controllers/groups.js#L915))
+ Set `invited` to false when:
  + Quest is cancelled before it starts (all group members; [code link](https://github.com/HabitRPG/habitrpg/blob/0fa93506f5ea2eb68284486dc0a37bd970c1e7a3/website/src/controllers/groups.js#L980-982))
  + User accepts the invitation (single user; [code link](https://github.com/HabitRPG/habitrpg/blob/0fa93506f5ea2eb68284486dc0a37bd970c1e7a3/website/src/controllers/groups.js#L951))
  + User rejects the invitation (single user; [code link](https://github.com/HabitRPG/habitrpg/blob/0fa93506f5ea2eb68284486dc0a37bd970c1e7a3/website/src/controllers/groups.js#L962))
  + Quest starts (all group members; [code link](https://github.com/HabitRPG/habitrpg/blob/0fa93506f5ea2eb68284486dc0a37bd970c1e7a3/website/src/models/group.js#L145), function called [here](https://github.com/HabitRPG/habitrpg/blob/0fa93506f5ea2eb68284486dc0a37bd970c1e7a3/website/src/controllers/groups.js#L835))
+ Replaced the quest invite modal notification logic with `user.party.quest.invited == true`, causing the invite modal to popup on user sync without causing overhead issues as discussed in [#5239](https://github.com/HabitRPG/habitrpg/pull/5239#issuecomment-103699467).

Everything is working as intended, except for the rewards display. The quest invite modal is currently displaying like this:

![selection_022](https://cloud.githubusercontent.com/assets/7059890/7828154/b7452faa-03fd-11e5-8fde-68c4a657bc1e.png)

...and I can't figure out why. I'm guessing it has something to do with the scope of the [`quest-rewards`](https://github.com/HabitRPG/habitrpg/blob/develop/common/script/public/directives.js#L155-164) directive, but I'm really not sure. pinging @crookedneighbor and @Alys for help.

### EDIT by Blade: Still needs to be done
- [ ] Migration script to add `party.quest.RSVPNeeded = true` to users who have pending quests
- [x] Tests around watch function

### Would be nice to do
- [x] Accept and reject buttons on party page (migration script less necessary)
- [x] Party sync when party page loads
- [x] ~~Refactor fetch recent messages~~ EDIT by hairlessbear: to be done in a later PR
- [x] Quest completion modal should appear before quest invite modal