@crookedneighbor 2015-05-27
I feel like I had the same problem, but I don't remember what I did to fix it. I'll play with it and get back to you.

In the meantime, something you've done has caused the push notification test for receiving a quest invite to fail. That'll need to be sorted out before this can be merged.