@crookedneighbor 2015-05-27
Oh I see what the test problem is regarding the test. I was mocking the User model in the test, and you added a call to user.update, which was not being mocked .