@crookedneighbor 2015-05-27
All that's left is to fix the api test:

Change userMock on line 92 of test/api/pushNotifications.coffee to:

```coffeescript
userMock = {
  findById: (arg, cb) ->
    cb(null, recipient)
  find: (arg, arg2, cb) ->
    cb(null, [recipient])
  update: (arg, arg2) ->
    { exec: -> true }
}
```

And it'd probably be good to write some tests around the watch, but that can be a tad confusing, so I'll write those wants this is merge ready.