@hairlessbear 2015-05-27
I figured out a hacky fix. I'm setting `user.party.quest.key` when an invitation is sent, and using that instead of `party.quest.key` when calling `quest-rewards` on the invite modal. 

Only one issue left now: when a user accepts or rejects the quest, if you refresh the page, the invite modal pops up again. But, if you refresh the page again, it doesn't show up. I'm guessing there's some sort of issue with `User.update`, but I'm not sure what yet. After `accept` or `reject` are clicked, the vlaues in mongoDB are being updated properly, so I guess they aren't being updated in the browser or something...? I'm investigating now.