@crookedneighbor 2015-05-27
Haven't, looked at it yet, but you may need to call user sync when the user makes a decision to force the browser to be in sync with the server. 

Probably what is happening is you refresh, the browser's version of the data says the RSVPNeeded is true and then the modal pops up. The next action is for the browser to sync with the server, which sets RSVPNeeded correctly. Just a theory. Does that make sense?

The trick will be to sync AFTER the server is done saving.