@hairlessbear 2015-05-27
The problem only occurs if the user refreshes immediately after clicking accept/reject without doing anything else. As soon as anything is done to cause user sync (e.g., check off a task), refreshing doesn't do anything. The easiest solution would be to force a user sync after each update call; I'll try to get that working.