@hairlessbear 2015-05-27
Almost. I updated all of the items in the invite modal to use `user.party.quest.key` for consistency, and changed the notification to not explicitly check for `true`. I'll push both right now.