@hairlessbear 2015-05-27
Oh, and I removed `$rootScope.party.$get` from the watch because it's no longer necessary (it was there for `party.quest.key` in the invite modal). Feel free to put it back if I was wrong in removing it.