@crookedneighbor 2015-05-28
@hairlessbear Well, the good news is I did get it to sync after calling party.$questAccept and $party.questReject by utilizing that those functions are promises, so I can call User.sync() after the request to the server resolves: https://github.com/HabitRPG/habitrpg/commit/42b482cb7efa8b37941e9fa0a58bba7a7bf3a4c6

Bad news: I'm not super great at promises and I don't know how to effectively mock and test them. Will do some research about how best to do that. 

I'd also like to explore attaching that User.sync() function on the end of all these promises: https://github.com/HabitRPG/habitrpg/blob/develop/website/public/js/services/groupServices.js#L34-L37 which would eliminate the need to create wrapper functions. Not sure if that's easy to do ,or even possible, in the current framework.