@hairlessbear 2015-05-28
I've never heard of promises, so I'm off to Google to learn. I'm glad you got it working though!

We definitely need to figure out a way to sync on `questCancel`. I confirmed via testing that the same refresh issue occurs if a quest is cancelled. It only happens when multiple actions are performed by multiple users in a specific order, so it would likely almost never occur in production, but obviously we still need to fix it.