@crookedneighbor 2015-05-29
FYI, this or would have been merged in automatically when the other one gets merged as it contains all the commits in this one. GitHub Magic!