@hairlessbear 2015-05-29
Does that mean I need to reopen it, or will the magic happen regardless?