@crookedneighbor 2015-05-29
Happens regardless. I should have just opened a pr into your branch, rather than creating a new PR. wasn't aware that I could do that until today.