@crookedneighbor 2015-05-29
Following questions are from @hairlessbear:

>Regarding making the party page fetch on load, would #4816 be a problem? I'm worried that those in large parties would never see the invitation statuses of some members.

There should be a difference between fetching the party and fetching the recent messages. Might be a good time to revisit the fetch recent messages code and see if it needs to be adjusted.

>Regarding putting the accept/reject buttons on the party page, if #5157 is going to be merged, we should probably add that functionality there instead of here, right?

I don't think it necessarily needs to go there, but adding it here would mitigate the need for a migration script, since the few users that get a quest request in between deploys would still be able to join the quest from the party page. 

>Also, since you've now written the tests that you asked for in #5239, should that behavior be lumped into this PR?

I'm not quite happy with where that feature is at, so let's leave it separate for now.