@crookedneighbor 2015-05-29
Actually, I've changed my mind. While we're cleaning up all these things, let's add #5239 (make quest completion modal appear before quest invite modal) to the list of nice to haves (see original post at top of PR)