@hairlessbear 2015-05-29
>There should be a difference between fetching the party and fetching the recent messages. Might be a good time to revisit the fetch recent messages code and see if it needs to be adjusted.

I don't think there is a difference currently (I may be misunderstanding the code). `Fetch recent messages` calls [`sync(group)`](https://github.com/HabitRPG/habitrpg/blob/develop/website/views/options/social/chat-box.jade#L22), which I'm assuming means the whole party is fetched.

>I don't think it necessarily needs to go there, but adding it here would mitigate the need for a migration script, since the few users that get a quest request in between deploys would still be able to join the quest from the party page.

Avoiding the migration script is something I hadn't thought of; that's a good point. I suppose we could add accept/reject buttons to this PR, then once it's merged, I could rebase the other PR and integrate the new buttons. On the other hand, if #5157 gets merged first, we could rebase here and add the buttons. Both options require approximately the same amount of effort.