@hairlessbear 2015-05-29
Regarding the `fetch recent messages` issue, [it looks like someone started working on it a while ago](https://github.com/HabitRPG/habitrpg/blob/develop/website/public/js/services/groupServices.js#L75-90), but left it as a "to-do".