@crookedneighbor 2015-05-30
@hairlessbear Good work on making the quest completion modal appear before the next invite modal. :+1: Works really well now.