@hairlessbear 2015-05-30
Thanks!

Would you prefer that a migration script be written for this, or the accept/reject buttons added to the party page? Only one is necessary, but both are currently in the checklists at the top.