@crookedneighbor 2015-06-11
Do the accept/reject buttons. I'll work with Alys to do the migration script.