@Alys 2015-06-16
For the migration script, if I understand correctly, you want to find all users who exist in a party's `quest.members` object with a value of `null` (they have not accepted nor rejected the quest), and you want to give those players the `party.quest.RSVPNeeded` attribute with a value of true.

If that's not correct, ignore everything below. :)

For this kind of migration, by far the easiest way to do it is to capture all quest.members for all parties with inactive quests into an `output` file (script takes about 10 to 20 seconds to run on the production database):
````
db.groups.find( {
    'type': 'party',
    'quest.active': false,
    'quest.key': {
        $exists:true,
        $ne:null,
        $ne:"",
    },
}, {
    'quest.members':1
}).pretty();
````

`output` then contains data like this:
````
{
    "_id" : "12340a62-4a34-4cd5-affe-3e11099a6082",
    "quest" : {
        "members" : {
            "1234fd38-9470-4893-84e0-0c62cd9b4ef7" : true,
            "1234a067-32d8-4be5-8bb8-eb4e526ad8ed" : null
        },
    },
}
{
  ... next party ....
}
````

Then use grep to get the members with a value of null, use sed to strip everything from those lines that isn't a UUID, and save the UUIDs to a temp file:
`cat output | grep null | sed -e 's/" : .*/",/' > /tmp/t`

then read the temp file into a script that sets the required attribute (script takes 10 to 20 seconds on a local copy of the production database, and probably slightly longer on the real one):
````
var uuidArray=[
  ... UUIDS go here ...
];
db.users.update({ '_id':{$in: uuidArray} }, {$set: {
	'party.quest.RSVPNeeded': true
}},{ multi: true });
````

The whole process is done in under a minute. 

You risk incorrectly setting party.quest.RSVPNeeded to true for any players who accept a quest within that minute, but it's a low risk, and we can check for that having happened afterwards with another couple of database queries. Even with that final check, this is much faster than writing, testing, and debugging an all-in-one script.